requirements.txt
feedparser
requests
openai
anthropic
pydub
audioop-lts
jinja2
python-dateutil
google-api-python-client
google-auth
google-auth-oauthlib
